+++ src/r0.rs
@@ -10,6 +10,7 @@ pub mod context;
 pub mod device;
 pub mod directory;
 pub mod filter;
+pub mod keys;
 pub mod media;
 pub mod membership;
 pub mod message;