+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Improvements:
 * Add `r0::room::get_room_event` (introduced in r0.4.0)
 * Add `r0::read_marker::set_read_marker` (introduced in r0.4.0)
 * Add `r0::capabilities::get_capabilities` (introduced in r0.5.0)
+* Add `r0::keys` endpoints (introduced in r0.3.0)
 
 # 0.5.0